Fix #19 Introduce CRSToCRS class
In newer versions of proj, you need to create a crs_to_crs projection and then transform coordinates with that, whereas in older versions you would pass 2 separate CRS's into the transform function.

This PR:
- Create a new class RGeo::CoordSys::CRSToCRS that is instantiated from 2 RGeo::CoordSys::Proj4 objects and stores the crs_to_crs.
- Create a singleton class RGeo::CoordSys::CRSStore that maps some combination of Proj4 objects to a CRSToCRS object.
When running transform_coords, check the CRSStore to see if the mapping exists, if not create it and store it, else get it.
- Change the underlying C function to accept the crs_to_crs instead of 2 projections and just transform the coordinates